backend/requirements_dev.txt
asgiref==3.8.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
cachecontrol==0.14.0
cachetools==5.5.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
Click==7.0
cloud-init==24.3.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==43.0.3
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
dj-database-url==2.3.0
Django==4.2
django-environ==0.11.2
entrypoints==0.3
firebase-admin==6.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==1.1.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.3
pexpect==4.6.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.22
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.7.2
sqlparse==0.5.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
djangorestframework>=3.15
Markdown==3.3.7
django-filter==21.1
Pillow
